make a new go module dedicated for fabric-ibc-lightclientd
The main go module uses ibc-go@v4 and ibc-go@v4 depends on
cosmos-sdk@v0.45. However, fabric-ibc-lightclientd can't use this
version of cosmos-sdk because yui-fabric-ibc is not compatible with this
version of cosmos-sdk (<=v0.43 is OK).

To solve this conflict, this commit adds a new module dedicated for
fabric-ibc-lightclientd. The new module uses ibc-go@v1.0.0-beta1 and
cosmos-sdk@v0.43.0-beta1.

Signed-off-by: Masanori Yoshida <masanori.yoshida@datachain.jp>